pack: Move packing procedure from main to State::pack
